⚡ Bolt: Centralize usePrefersReducedMotion hook to reduce duplication
Extract the `prefers-reduced-motion` logic into a shared `usePrefersReducedMotion` hook using `useSyncExternalStore`.
Refactor 9 components (Button, LoadingSpinner, Tooltip, ScrollToTop, Skeleton, ToastContainer, BlueprintDisplay, StepCelebration, SuccessCelebration) to use this centralized hook.
This reduces code duplication, decreases bundle size by ~100 lines, and ensures consistent accessibility behavior across the application.
